+++ Budget-API/Budget.Models/DTO/SettingsDTO.cs
@@ -4,7 +4,7 @@ namespace Budget.Models.DTO
 {
     public class SettingsDTO
     { 
-        public long BudgetAmount { get; set; }
+        public double BudgetAmount { get; set; }
         public int Day { get; set; }
     }
 }
